Add files via upload
use this to force retrobat to run the ps1 file on startup. 

my main admin account only needs the ps1 file. but when running retrobat from my user profile in a windows sandbox environment retrobat refuses to run the script, with this bat file present it forces retrobat to run the ps1 regardless.